firstcommit

+++ database/migrations/2020_03_08_203937_create_products_table.php
@@ -16,11 +16,11 @@ class CreateProductsTable extends Migration {
             $table->id();
             $table->string('image');
             $table->integer('popularity')->nullable();
-            $table->integer('price');
+            $table->decimal('price', 6, 2);
             $table->integer('ratings')->nullable();
             $table->boolean('is_New')->nullable();
             $table->boolean('is_Discount')->nullable();
-            $table->integer('Discount')->nullable();
+            $table->decimal('Discount', 2, 2)->nullable();
             $table->bigInteger('sub_subcategory_id')->unsigned()->nullable();
             $table->foreign('sub_subcategory_id')->references('id')->on('sub_sub_categories')->onDelete('cascade');
             $table->timestamps();
